requirements.txt
flask==1.0.2
baidu-aip==2.2.11.0
pytest==4.1.1
coverage==4.5.2
pytest-cov==2.6.1
sounddevice==0.3.12
numpy==1.16.0